refactor(auth): Split domain and database infra modules
