+++ auth/src/domain/mod.rs
@@ -1,7 +1,3 @@
-// Migration
-pub mod migration;
-
-// Entity
 pub mod permission;
 pub mod role;
 pub mod role_permissions;
@@ -9,8 +5,3 @@ pub mod user_info;
 pub mod user_role;
 pub mod user_settings;
 pub mod users;
-
-pub mod common;
-mod init;
-
-pub use init::*;
