+++ auth/src/domain/user_info.rs
