+++ auth/src/features/api/v1/auth/verify_email.rs
@@ -5,8 +5,8 @@ use serde::Deserialize;
 use token::{TokenStore, services::RegisterTokenService};
 
 use crate::{
+    domain::users,
     internal::error::{AppError, AppErrorKind},
-    models::users,
     routers::{
         extractor::Json,
         response::app_error_to_response,
