+++ auth/src/infra/database/error.rs
