+++ auth/src/infra/database/migration/m20241010_000000_create_table.rs
