+++ auth/src/infra/database/migration/m20241201_000000_add_userinfo.rs
