+++ auth/src/infra/database/migration/m20260311_000000_create_user_settings.rs
