+++ auth/src/infra/database/mod.rs
@@ -1,0 +1,7 @@
+pub mod migration;
+
+mod error;
+mod init;
+
+pub use error::*;
+pub use init::*;
