+++ auth/src/infra/mod.rs
@@ -1,0 +1,1 @@
+pub mod database;
